Separate into a sibling package instead of being a child of pulse.
