+++ src/main/java/io/pivotal/pulse/CreateProjectController.java
@@ -1,6 +1,6 @@
 package io.pivotal.pulse;
 
-import io.pivotal.pulse.domain.CreateProjectUseCase;
+import io.pivotal.pulsedomain.CreateProjectUseCase;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
