--- src/main/java/io/pivotal/pulse/domain/ProjectEntity.java
@@ -1,4 +1,0 @@
-package io.pivotal.pulse.domain;
-
-public class ProjectEntity {
-}
