+++ src/main/java/io/pivotal/pulsedomain/CreateProjectUseCase.java
@@ -1,4 +1,5 @@
-package io.pivotal.pulse.domain;
+package io.pivotal.pulsedomain;
+
 
 public class CreateProjectUseCase {
