+++ src/main/java/io/pivotal/pulsedomain/ProjectEntity.java
@@ -1,0 +1,4 @@
+package io.pivotal.pulsedomain;
+
+public class ProjectEntity {
+}
