+++ src/main/java/io/pivotal/pulsedomain/ProjectRepository.java
@@ -1,4 +1,4 @@
-package io.pivotal.pulse.domain;
+package io.pivotal.pulsedomain;
 
 public interface ProjectRepository {
 }
